Added unit tests

+++ src/main/java/org/graylog/plugins/aggregates/rule/Rule.java
@@ -8,7 +8,7 @@ public interface Rule {
 
 	public String getField();
 
-	public int getNumberOfMatches();
+	public long getNumberOfMatches();
 
 	public boolean isMatchMoreOrEqual();
